Add working AppContext and provider through all the app
