+++ src/context/index.ts
@@ -1,0 +1,5 @@
+import { createContext } from "react";
+
+const AppContext = createContext({});
+
+export { AppContext };
