+++ src/main.tsx
@@ -1,30 +1,17 @@
 import React from "react";
 import ReactDOM from "react-dom/client";
-import App from "./App.tsx";
-import { BrowserRouter, useRoutes } from "react-router-dom";
-import { ProfilePage } from "./pages/profile.tsx";
+import { BrowserRouter } from "react-router-dom";
 import { AsideNav } from "./components/AsideNav/index.tsx";
-
-function AppRoutes() {
-  const routes = useRoutes([
-    {
-      // TODO: Añadir una página de error personalizada con la propiedad de errorPage.
-      path: "/",
-      element: <App />,
-    },
-    {
-      path: "/profile",
-      element: <ProfilePage />,
-    },
-  ]);
-  return routes;
-}
+import { AppProvider } from "./context/AppProvider.tsx";
+import { AppRoutes } from "./AppRoutes.tsx";
 
 ReactDOM.createRoot(document.getElementById("root")!).render(
   <React.StrictMode>
-    <BrowserRouter>
-      <AsideNav />
-      <AppRoutes />
-    </BrowserRouter>
+    <AppProvider>
+      <BrowserRouter>
+        <AsideNav />
+        <AppRoutes />
+      </BrowserRouter>
+    </AppProvider>
   </React.StrictMode>,
 );
