--- src/utils/currentDate.ts
@@ -1,2 +1,0 @@
-const now = new Date();
-export const currentDate = now.toLocaleDateString();
